- 1 April 2021 new merge

+++ src/main/java/com/psikku/backend/repository/FileDataRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface FileDataRepository extends JpaRepository<FileData,Integer> {
+public interface FileDataRepository extends JpaRepository<FileData,Long> {
 
     Optional<FileData> findByFileName(String fileName);
     Optional<FileData> findByFilePathAndFileName(String path,String fileName);
